+++ packages/app/src/mw/index.ts
@@ -1,4 +1,0 @@
-export * from './cors';
-export * from './serve';
-export * from './jwt';
-export * from './view';