+++ packages/core/src/server.ts
@@ -59,6 +59,11 @@ export class RpcServer {
   private plugins: Strategy[] = [];
   app;
   use;
+  proxy = {
+    init: [],
+    load: [],
+    before: {},
+  };
   config;
   init: [];
   load: [];
@@ -117,7 +122,28 @@ export class RpcServer {
       await next();
     });
 
+    // proxy
+    for (const plugin of this.plugins) {
+      console.dir('init proxy stage');
+      if (plugin.config.proxy) {
+        if (plugin.config.proxy.inject === 'init') {
+          this.proxy.init.push(plugin.proxy());
+        }
+        if (plugin.config.proxy.inject === 'load') {
+          this.proxy.load.push(plugin.proxy());
+        }
+        if (plugin.config.proxy.inject === 'before') {
+          for (const i in plugin.config.proxy.before) {
+            const name = plugin.config.proxy.before[i];
+            if (!this.proxy.before[name.toLowerCase()]) this.proxy.before[name.toLowerCase()] = [];
+            this.proxy.before[name.toLowerCase()].push(plugin.proxy());
+          }
+        }
+      }
+    }
+
     // init
+    this.config.hooks.init = this.proxy.init;
     for (const plugin of this.plugins) {
       console.dir('init stage');
       if (plugin.init.length > 0) this.config.hooks.init.push(...plugin.init);
@@ -129,6 +155,7 @@ export class RpcServer {
     this.config.init(this);
 
     // load
+    this.config.hooks.load = this.proxy.load;
     for (const plugin of this.plugins) {
       console.dir('load stage');
       if (plugin.load.length > 0) this.config.hooks.load.push(...plugin.load);
@@ -144,14 +171,12 @@ export class RpcServer {
     // mount app
     for (const plugin of this.plugins) {
       console.dir('mount plugin ' + plugin.prefix);
-      // console.dir(plugin);
-      // this.app.use(compose([plugin.proxy(), mount(plugin.prefix, plugin.app)]));
 
-      // console.dir('prefix is string ' + plugin.prefix);
+      const mw = this.proxy.before[plugin.name.toLowerCase()] || [];
 
-      const mw = plugin.prefix === '' ? mount(plugin.app) : mount(plugin.prefix, plugin.app);
+      const app = plugin.prefix === '' ? mount(plugin.app) : mount(plugin.prefix, plugin.app);
       console.dir(plugin.prefix);
-      this.app.use(mw);
+      this.app.use(compose([...mw, app]));
 
       // console.dir(plugin.prefix === '');
       // console.dir(mw);