+++ packages/app/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @tomrpc/mount
 
+## 2.0.0
+
+### Major Changes
+
+- test
+
+### Patch Changes
+
+- Updated dependencies
+  - @tomrpc/client@2.0.0
+  - @tomrpc/mount@2.0.0
+  - @tomrpc/core@2.0.0
+
 ## 1.0.0
 
 ### Major Changes