+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tomrpc/core
 
+## 2.0.0
+
+### Major Changes
+
+- test
+
 ## 1.0.0
 
 ### Major Changes